requirements.txt
# Ingestion APIs
requests==2.33.1
pandas==3.0.2

# Tests
pytest==9.0.2

# Variables d'environnements
python-dotenv==1.2.2

#Google Cloud Storage
google-cloud-storage==3.10.1
